+++ src/Symfony/Bundle/FrameworkBundle/Tests/Functional/SchedulerTest.php
@@ -12,29 +12,29 @@
 namespace Symfony\Bundle\FrameworkBundle\Tests\Functional;
 
 use Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Messenger\BarMessage;
-use Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Messenger\DummyScheduleConfigLocator;
+use Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Messenger\DummySchedule;
 use Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Messenger\FooMessage;
 use Symfony\Component\Clock\MockClock;
 use Symfony\Component\HttpKernel\KernelInterface;
-use Symfony\Component\Scheduler\Messenger\ScheduleTransport;
-use Symfony\Component\Scheduler\Schedule\ScheduleConfig;
+use Symfony\Component\Scheduler\Messenger\SchedulerTransport;
+use Symfony\Component\Scheduler\RecurringMessage;
 use Symfony\Component\Scheduler\Trigger\PeriodicalTrigger;
 
 class SchedulerTest extends AbstractWebTestCase
 {
     public function testScheduler()
     {
-        $scheduleConfig = new ScheduleConfig([
-            [PeriodicalTrigger::create(600, '2020-01-01T00:00:00Z'), $foo = new FooMessage()],
-            [PeriodicalTrigger::create(600, '2020-01-01T00:01:00Z'), $bar = new BarMessage()],
-        ]);
-        DummyScheduleConfigLocator::$schedules = ['default' => $scheduleConfig];
+        $scheduledMessages = [
+            new RecurringMessage($foo = new FooMessage(), PeriodicalTrigger::create(600, '2020-01-01T00:00:00Z')),
+            new RecurringMessage($bar = new BarMessage(), PeriodicalTrigger::create(600, '2020-01-01T00:01:00Z')),
+        ];
+        DummySchedule::$recurringMessages = $scheduledMessages;
 
         $container = self::getContainer();
         $container->set('clock', $clock = new MockClock('2020-01-01T00:09:59Z'));
 
-        $this->assertTrue($container->get('receivers')->has('schedule'));
-        $this->assertInstanceOf(ScheduleTransport::class, $cron = $container->get('receivers')->get('schedule'));
+        $this->assertTrue($container->get('receivers')->has('scheduler_dummy'));
+        $this->assertInstanceOf(SchedulerTransport::class, $cron = $container->get('receivers')->get('scheduler_dummy'));
 
         $fetchMessages = static function (float $sleep) use ($clock, $cron) {
             if (0 < $sleep) {
@@ -42,7 +42,7 @@ class SchedulerTest extends AbstractWebTestCase
             }
             $messages = [];
             foreach ($cron->get() as $key => $envelope) {
-                $messages[$key] = $envelope->getMessage();
+                $messages[$key] = $envelope->getMessage()->envelope->getMessage();
             }
 
             return $messages;