+++ src/Symfony/Component/Scheduler/Tests/Messenger/SchedulerTransportTest.php
@@ -13,12 +13,13 @@ namespace Symfony\Component\Scheduler\Tests\Messenger;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Messenger\Envelope;
-use Symfony\Component\Scheduler\Exception\LogicMessengerException;
+use Symfony\Component\Messenger\Message\RedispatchMessage;
+use Symfony\Component\Scheduler\Exception\LogicException;
+use Symfony\Component\Scheduler\Generator\MessageGeneratorInterface;
 use Symfony\Component\Scheduler\Messenger\ScheduledStamp;
-use Symfony\Component\Scheduler\Messenger\ScheduleTransport;
-use Symfony\Component\Scheduler\Schedule\ScheduleInterface;
+use Symfony\Component\Scheduler\Messenger\SchedulerTransport;
 
-class ScheduleTransportTest extends TestCase
+class SchedulerTransportTest extends TestCase
 {
     public function testGetFromIterator()
     {
@@ -26,13 +27,15 @@ class ScheduleTransportTest extends TestCase
             (object) ['id' => 'first'],
             (object) ['id' => 'second'],
         ];
-        $scheduler = $this->createConfiguredMock(ScheduleInterface::class, [
+        $generator = $this->createConfiguredMock(MessageGeneratorInterface::class, [
             'getMessages' => $messages,
         ]);
-        $transport = new ScheduleTransport($scheduler);
+        $transport = new SchedulerTransport($generator);
 
         foreach ($transport->get() as $envelope) {
             $this->assertInstanceOf(Envelope::class, $envelope);
+            $this->assertInstanceOf(RedispatchMessage::class, $envelope->getMessage());
+            $envelope = $envelope->getMessage()->envelope;
             $this->assertNotNull($envelope->last(ScheduledStamp::class));
             $this->assertSame(array_shift($messages), $envelope->getMessage());
         }
@@ -42,7 +45,7 @@ class ScheduleTransportTest extends TestCase
 
     public function testAckIgnored()
     {
-        $transport = new ScheduleTransport($this->createMock(ScheduleInterface::class));
+        $transport = new SchedulerTransport($this->createMock(MessageGeneratorInterface::class));
 
         $transport->ack(new Envelope(new \stdClass()));
 
@@ -51,17 +54,17 @@ class ScheduleTransportTest extends TestCase
 
     public function testRejectException()
     {
-        $transport = new ScheduleTransport($this->createMock(ScheduleInterface::class));
+        $transport = new SchedulerTransport($this->createMock(MessageGeneratorInterface::class));
 
-        $this->expectException(LogicMessengerException::class);
+        $this->expectException(LogicException::class);
         $transport->reject(new Envelope(new \stdClass()));
     }
 
     public function testSendException()
     {
-        $transport = new ScheduleTransport($this->createMock(ScheduleInterface::class));
+        $transport = new SchedulerTransport($this->createMock(MessageGeneratorInterface::class));
 
-        $this->expectException(LogicMessengerException::class);
+        $this->expectException(LogicException::class);
         $transport->send(new Envelope(new \stdClass()));
     }
 }